+++ lib/event_store/registration/local_registry.ex
@@ -1,61 +1,62 @@
 defmodule EventStore.Registration.LocalRegistry do
   @moduledoc """
-  Local process registration, restricted to a single node, using Elixir's [Registry](https://hexdocs.pm/elixir/Registry.html)
+  Pub/sub using Elixir's local `Registry` module, restricted to running on a
+  single node only.
   """
 
   @behaviour EventStore.Registration
 
   @doc """
-  Return the local supervisor child spec
+  Return the local supervisor child spec.
   """
   @spec child_spec() :: [:supervisor.child_spec()]
   @impl EventStore.Registration
   def child_spec do
     [
-      Supervisor.child_spec({Registry, keys: :unique, name: EventStore.Registration.LocalRegistry}, id: EventStore.Registration.LocalRegistry),
+      Supervisor.child_spec(
+        {
+          Registry,
+          keys: :duplicate, name: EventStore.PubSub, partitions: System.schedulers_online()
+        },
+        id: EventStore.PubSub
+      )
     ]
   end
 
   @doc """
-  Starts a uniquely named child process of a supervisor using the given module and args.
-
-  Registers the pid with the given name.
+  Subscribes the caller to the given topic.
   """
-  @spec start_child(name :: term(), supervisor :: module(), args :: [any()]) :: {:ok, pid()} | {:error, reason :: term()}
+  @spec subscribe(binary) :: :ok | {:error, term}
   @impl EventStore.Registration
-  def start_child(name, supervisor, args) do
-    case whereis_name(name) do
-      :undefined ->
-        via_name = {:via, Registry, {EventStore.Registration.LocalRegistry, name}}
-
-        Supervisor.start_child(supervisor, args ++ [[name: via_name]])
-
-      pid ->
-        {:ok, pid}
+  def subscribe(topic) do
+    with {:ok, _} <- Registry.register(EventStore.PubSub, topic, []) do
+      :ok
     end
   end
 
   @doc """
-  Sends a message to the given dest and returns `:ok`.
+  Is the caller subscribed to the given topic?
   """
-  @callback multi_send(dest :: atom(), message :: any()) :: :ok
+  @spec subscribed?(binary) :: true | false
   @impl EventStore.Registration
-  def multi_send(server, message) do
-    send(server, message)
-    :ok
+  def subscribed?(topic) do
+    subscriptions = Registry.lookup(EventStore.PubSub, topic)
+    caller = self()
+
+    Enum.any?(subscriptions, fn
+      {^caller, _} -> true
+      _ -> false
+    end)
   end
 
   @doc """
-  Get the pid of a registered name.
+  Broadcasts message on given topic.
   """
-  @spec whereis_name(name :: term) :: pid | :undefined
+  @spec broadcast(binary, term) :: :ok | {:error, term}
   @impl EventStore.Registration
-  def whereis_name(name), do: Registry.whereis_name({EventStore.Registration.LocalRegistry, name})
-
-  @doc """
-  Return a `:via` tuple to route a message to a process by its registered name
-  """
-  @spec via_tuple(name :: term()) :: {:via, module(), name :: term()}
-  @impl EventStore.Registration
-  def via_tuple(name), do: {:via, Registry, {EventStore.Registration.LocalRegistry, name}}
+  def broadcast(topic, message) do
+    Registry.dispatch(EventStore.PubSub, topic, fn entries ->
+      for {pid, _} <- entries, do: send(pid, message)
+    end)
+  end
 end